#866 delete unnecessary stuff
